LU-123 Port yaml and auster to b1_8
Changes to add the yaml data logging from the 2.0 branch in the 1.8 branch, this
patch was created by applying the 2.0 yml patch to 1.8 and then resolving the issues.

This is to say that the larger changes have been taken to be correct because of their
acceptance in the 2.0 master.

This patch also contains the addition of the auster.sh script which allows for the logging
of test results to the maloo database. The actual logging is carried out by the
test-results.rb script.

For instructions on how to use auster use auster -?, for instructions on maloo refer to
the whamcloud wiki [where I will begin to place a resource of information!]

Change-Id: I602a3534f17544d857aa0a9f9f82d2873fb73a39
Signed-off-by: Chris Gearing <chris@whamcloud.com>
Signed-off-by: Bobi Jam <bobijam@whamcloud.com>