LU-120 the perfect storm
There is a corner case in the debian patching/unpatching process
where a patch converts the use of a given file to a file.in that is
to be substed.  The problem is that between the configure and the
make debs, a file, which a debian patch wants to remove already
exists (it was created by the configure) and dpkg-buildpackage's
unpatch step fails becuase it wants to try to create a file which
configure already created.

Signed-off-by: Brian J. Murrell <brian@whamcloud.com>
Change-Id: Ibb06c78e48cfb56cb35d1226f166a5740114661f