LU-62 lbuild and friends for RHEL6 patchless client
This adds the lbuild support needed to build RHEL6 patchless
client.
Included is a re-org of the RHEL5 lbuild code to maximize code
reuse and minimze code replication.
The bits to build the RHEL6 kernel are not actually needed but
included just to keep parity with the master branch where RHEL6
kernel and server is built and supported.

Signed-off-by: Brian J. Murrell <brian@whamcloud.com>
Change-Id: Ie5ff1249aee8725625901d1dad242a8f8c48765a