b=18750 Make failure handling and injection common to all
This patch moves the failure injection system lustre uses
for debugging into the libcfs layer so other sub systems
of lustre can take advantage of them. For example a LNET
driver could set the fail_loc to test the behavior of the
LNET layer for the case of a particular hardware failure.

Issue: LU-65
Change-Id: Ib7b9cf80a768bd41e72f5bb046f089351fb69073
Signed-off-by: James Simmons <uja.ornl@gmail.com>