LU-169 implement layout lock for LOV EA
this patch:
- implements a new inode lock bit for layout lock
- adds lsm refcounting
- adds a generation number to the LOV EA
Each time a lsm is used, a reference must be taken with
ll_lsm_get() or ll_lsm_get_io().
A layout lock is acquired when taking a reference on a lsm.
If the lsm is needed for an io, the lock is hold up to the extent lock
acquisition. If the lsm is not needed for io, the lock is given back
just after reference is taken.
When lsm is no more needed, if must be unreferenced with ll_lsm_put().

Change-Id: I23b9bbfeeb6a59bbb21a2725bc2ee4d39ce14084
Signed-off-by: jcl <jacques-charles.lafoucriere@cea.fr>