LU-164 Update testframework to restart sessions at next suite
Update the test framework so that it can append to yaml files rather
than begin a new series and allow auster to begin at a selected
test_suite.

auster has a new switch '-S suite-name' which allows the caller to
start a test group part way through.

The idea behind all of this is that a test session that fails at a
test suite can be restarted at the next suite.

Signed-off-by: Chris Gearing <chris@whamcloud.com>
Change-Id: Ieaa987874d69fa8c33c3af92957e99607a8a7be7