LU-202 fix the test_23 of sanity-quota.sh
The test_23() of sanity-quota.sh uses a single stripe file to do the test,
however, the test script always write (OSTCOUNT * N)M bytes to the test file,
it doesn't make sense at all.

Moreover, test_23 uses a constant value(120000 = ~120M) to check if the
system has enough space to finish the test, so in the large OSTCOUNT testing
environment, the test will fail for short of disk space.

Changes of this patch are:
- Always Write N bytes to the test file instead of (OSTCOUNT * N);
- Use a reasonable value to check free space instead of magic 120000;

Issue:
Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: If7cd01b7ce9dc025a0f1b8b70fa9214d70182a86