LU-215 Bind ost/mdt to specific nids
The goal of the patch attached to this bug is to make NUMIOA factor
management easier, by giving the ability to restrict the NIDs that a
target (mdt or ost) registers on the MGS. By this means we can enforce
client requests to go through the most appropriate network interface,
ie the one that minimizes the NUMIOA factor.

Binding is controlled via the new '--network' option of mkfs.lustre and
tunefs.lustre commands.
This option takes as values one or more LNET networks. At mount.lustre
time the target registers on the MGS with only the NIDs that pertain to the
networks specified. If no specific network is given, all NIDs are registered.

Change-Id: I8f5e8ac71dd0d81948af11edc58572001207c37d
Bugzilla: 22078
Signed-off-by: Liang Zhen <liang@whamcloud.com>