LU-225 --with-ldiskfs-inkernel checks wrong file
Commit fd0bd149e66cbc06a17c372e159f9bfba024cc97 changed the file checked
when the --with-ldiskfs-inkernel configure option is used from
$LINUX/fs/ldiskfs/inode.c to $LINUX/include/linux/ldiskfs_fs.h.

It looks like the person fixing a similar case statement assumed the
same fix applied to the inkernel case.  However, this is wrong because
the purpose of --with-ldiskfs-inkernel is to allow us to build lustre
without having full kernel sources available, so inode.c won't exist.

Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: I6dc1db41fdee0016eab7d151c519e3359d78b44b