LU-227 External lustre-ldiskfs package integration
Add a configure option --with-ldiskfs-devel to build Lustre against the
externally maintained lustre-ldiskfs package. This external package maintains
ldiskfs as a standalone codebase, completely separate from Lustre source tree.
The intent is to eventually have a clear distinction between the Lustre package
and the package Lustre uses for it's underlying disk format.

Signed-off-by: Ned Bass <bass6@llnl.gov>
Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Iaa17d0483e155439c8de206af81f19a321bbeaa1