LU-224 Move fail_loc handling from lustre to libcfs
This patch goal is to move the fail_loc handling from lustre to libcfs. The
reason being is that currently you can only do fault testing in the lustre
core code. By moving this code to libcfs all sub systems will be able to inject
faults into the system for testing purposes. This is in particular needed to
support the cray gemini klnd driver since one can inject faults for testing
failure conditions.

Change-Id: I438c30d0c7fe6987ad3fcfa5442366d4307a4d9f
Signed-off-by: James Simmons <uja.ornl@gmail.com>