LU-251 Fix gcc configure warnings
Newer versions of gcc are getting smart enough to detect the sloppy
syntax used for the autoconf tests.  It is now generating warnings for
unused or uninitialized variables.  Newer versions of gcc even have the
-Wunused-but-set-variable option set by default.  This isn't a problem
except when -Werror is set and they get promoted to an error.  In this
case the autoconf test will return an incorrect result which will result
in a build failure or runtime error later on.

To handle this I'm tightening up many of the autoconf tests to
explicitly mark variables as unused to suppress the gcc warning.  Tests
emitting uninitialized variable errors are updated to initialize
pointers to NULL, and some variables are converted to pointers to
accomodate this.  'Argument makes integer from pointer without a cast'
errors were fixed by passing 0 for the offending argument in cases where
we are not explicitly testing the argument type.  0 is accepted as both
an integer and a pointer.

Change-Id: Idaa04b04308e3cd994b0d802a5ee1eb5c90f9be6
Signed-off-by: Ned Bass <bass6@llnl.gov>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Reviewed-on: http://review.whamcloud.com/478
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>