LU-202 fix the test_23 of sanity-quota.sh
The test_23() of sanity-quota.sh uses a single stripe file to do the test,
however, the test script always write (OSTCOUNT * N)M bytes to the test file,
it doesn't make sense at all.

Moreover, test_23 uses a constant value(120000 = ~120M) to check if the
system has enough space to finish the test, so in the large OSTCOUNT testing
environment, the test will fail for short of disk space.

Changes of this patch are:
- Always Write N bytes to the test file instead of (OSTCOUNT * N);
- Use a reasonable value to check free space instead of magic 120000;
- Specify the stripe index of test file for test_9/10 and test_23;

Issue:
Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: I2d9059d23cf53fee0244a226622b0e9adb800599