LU-118 Workaround to avoid race causing assertion failure in clear_inode()
This patch is just a *workaround* to avoid a race found in recent kernel
causing the following assertion in clear_inode(): BUG_ON(inode->i_data.nrpages).
The race can still occur when truncate_inode_pages() & clear_inode() are
called directly by the VFS, but the problem should be harder to reproduce.
A more complete fix is considered for 2.x.

Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Change-Id: I2389438440dd531caf23de69d876979ddd3e195c