LU-249 Allow the user to set the upper limit to the thread count on the MDS
Currently the largest number of the threads that can exist on the MDS is 512.
For very large system is those threads can easily be exhausted and end up back
logging the system. This patch allows the user to define the upper limit of
how many threads are created on the MDS. If not selected the default behavior
of 512 is keep.

Change-Id: I1952bbb5a446dd5debf5f787489e1c456da7945a
Signed-off-by: James Simmons <uja.ornl@gmail.com>