LU-258 Cleanup and reorganize obdfilter-survey
The goals of this patch:

1) Eliminate repetition.

2) Translate blocks of code to functions for increased readability.

3) Use more descriptive names for functions and variables.

4) Use consistent formatting

5) Update the documentation (README.obdfilter-survey)

The separate 'libecho' file is deleted because the case-specific code
is quite small. I don't think having separate scripts for each use-case
is particularly useful.

The script should be functionally equivalent to HEAD with one exception.
For the netdisk case HEAD will create local osc(s) if they do not
exist already. With the patch netdisk requires local osc(s).

The XXX: ? style comments are questions I had while doing the cleanup
and should not be in the final script.

Change-Id: I274353575666d0f423e52ca343410663d70c07ba
Signed-off-by: Chris Horn <hornc@cray.com>