LU-258 Cleanup and reorganize obdfilter-survey
The goals of this patch:

1) Eliminate repetition.

2) Translate blocks of code to functions for increased readability.

3) Use more descriptive names for functions and variables.

4) Use consistent formatting

5) Update the documentation (README.obdfilter-survey)

The separate 'libecho' file is deleted because the case-specific code
is quite small. I don't think having separate scripts for each use-case
is particularly useful.

The script should be functionally equivalent to HEAD with two
exceptions:
1) For the netdisk case HEAD will create local osc(s) if they do not
exist already. With the patch netdisk case requires existing local
osc(s).

2) For case=network use h2$NETTYPE functions to set the oss_nids value.
If, for example, we set targets="nid00037" then add_new_uuids will
fail with something like "can't parse lustre nid nid00037@gni.". This
patch uses the h2$NETTYPE functions from test-framework.sh to set the
oss_nids value. An alternative solution would be to define an OSSNID
variable which can be set at the command line or in the script. However,
I prefer the approach which doesn't require the user to do extra work.

The XXX: ? style comments are questions I had while doing the cleanup
and should not be in the final script.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: I274353575666d0f423e52ca343410663d70c07ba