LU-278 Don't abort configure if tag is of unrecognized form.
The configure process should NOT abort just because the most
recent tag is not of the form that upstream uses to tag Lustre.
Downstream developers may use their own tags, or just add
extensions to upsteam's version tags.

This patch is ugly, but at least it gets us building again.

Change-Id: I9a98bfd4475d3df2694f536ba0352779a62650c7
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>