LU-263 Fixup yaml.sh and test-framework.sh so that it only ever uses the first part of the host name for log files. This means stripping of everything off the name after and including the first .
This change is only designed to make the .yml files consistent. The log files
will append the fully qualified name and appear to do this consitantly.

We can create a jira to make the log and yml files consistant, but the reality
is that the yml files have a short life and are only used to send to Maloo.

This change will allow automated posting of results and automated testing to
begin.

1. Carries out the above using sed
2. Adds in a quick change so that LUSTRE_BUILD in yaml.sh can be a reference to
the source rather than just the lustre version string which is recorded and
written to the yaml anyway as LUSTER_VERSION

Additionally a couple of other changes sneaked in.

1. Allows the review information to be applied to the yaml output file
by way of exporting the variable CODE_REVIEW_YAML to be a yaml description
for maloo

2. The addition of a couple of fixes to make the permissions for yaml
files be allow-all. This permissions are bracketed and so do not change any
other parts of the code.

Resubmmited for review without change to trigger a new hudson build
previous build failed for reasons unrelated to this patch.

Change-Id: I40520cd5ee113927186341155d8a5e16888779ac
Signed-off-by: Chris Gearing <chris@whamcloud.com>