LU-281 Autoconf improvements and rpm generation fixes
For the autoconf system you can over ride the defaults where the
software stack will be installed on an given system image. For
lustre this is not completely true. This patch always you to
define the location where the kernel modules and mount.lustre
can be. Also another issue addressed is the case where many
distros pull in extra dependencies that are not needed when
a rpm is generated. Those extra dependencies might not exist
in a system image. AutoReqProv can suppress this.

Change-Id: I49e8b3b9fcd8edc52f93075032da96e248b1056f
Signed-off-by: James Simmons <uja.ornl@gmail.com>