LU-64 Fix test conditions in stripe attr helpers
If $ACTUAL happened to be the empty string, the test to determine if an
error occurred would report a false pass. This patch fixes this
issue by prepending the character 'x' to the variables during the
comparison.

Also, when an error is reported, the variables are now wrapped in single
quotes. This makes the output more readable if one of the variables is
the empty string.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Idd6564c2a0e6880dd3cd84d2385316380280cab2