LU-164 Update testframework so that it can restart sessions at the next suite
Update the test framework so that it can append to yaml files rather than begin a
new series and allow auster to begin at a selected test_suite.

auster has a new switch '-S suite-name' which allows the caller to start a test
group part way through.

The idea behind all of this is that a test session that fails at a test suite can
be restarted at the next suite.

Change-Id: Ifc84184506cff5ee11e5dc570ce17baab4d439a4
Signed-off-by: Chris Gearing <chris@whamcloud.com>