LU-372 add $TMP/mgsactive to indicate the active combined MGS/MDS node
The current test-framework could not figure out the active MGS node
correctly while the MGS and MDS nodes are combined under a failover
configuration. We need add a $TMP/mgsactive file to indicate the
active MGS node like what we did for the MDS node.

In addition, "set_default_debug_nodes()" has a defect to check
whether "$nodes" contains the local node or not. The patch contains
a fix for that.

Signed-off-by: Yu Jian <yujian@whamcloud.com>
Change-Id: Iae49497601d09fef38968360d8543cc753009e85