LU-384 land bz22281 on b1_8
To facilitate distinctive naming, we need the work in bugzilla
22281 landed on b1_8.
Add an lbuild option to set the downstream-release.  Normally
this should not be exposed to/through lbuild as this was orig-
inally intended for downstream packages to add their own stamp
to the release string.

Include:
    b=24095 don't really need configure --with-linux

    Don't really need to have run configure --with-linux for
    version-tag.pl to work.

    i=wangyb
    i=cliff

Expand the list of debian/* files that should be included in the
dist tarball.
Don't try to [un]patch the local-env.sh.  It should never really
be included in a landed patch.

Signed-off-by: Brian J. Murrell <brian@whamcloud.com>
Change-Id: I82f4f9a556fd1d1fc51ef8930e2017798ff88706