LU-395 Fix obd_zombie_barrier()
The obd_zombie_barrier() is waiting for obd_zombie_export list is empty
and exit, but export may be not yet destroyed. This produces race between export
destroying and other cleanups like lut_fini()

The patch introduces zombies_in_destroy variable which is counter, and
obd_zombie_barrier is checking it.

Change-Id: I0fd4a1023efade1cb40d4c36e59169a25a525ed1
Signed-off-by: Mikhail Pershin <tappro@whamcloud.com>