LU-136 do not use mke2fs -E resize for large LUNs
The mo_device_size variable is in units of kB, so take this into
account correctly when computing the filesystem block count.
The "-E resize" option is being used to align metadata on 1MB IO
boundaries, but it does not work for devices larger than 2^32 blocks.

A better long-term solution is to fix mke2fs so that it does this
alignment itself based on the s_raid_stride value.

Change-Id: I2e4104b8ad490076bfaa6a49842124d79e3bf880
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>