LU-73 kernel: remove wrong jbd2 commit timer patch
The bug that this patch was originally fixing (incorrectly rounding
down the jiffies timer, which might cause it to be skipped) was
fixed in the upstream kernel by using round_jiffies_up(), which
always moves the timer into the future.

Remove this patch entirely, since it is no longer needed.

Change-Id: I3b4f30e0565cf64b95fc87babf1f54a9ab190754
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>