LU-409 stop loading the llite_lloop module for >= 2.6.32 kernels
The change of block device interface in kernel 2.6.32-131 caused
the llite_lloop module to fail. llite_lloop module is currently
unsupported so this patch introduces a flag (default false)
to enable loading of llite_lloop module.

It is hoped that llite_lloop module will receive attention, so
this patch provides simple way to re-enable the llite_lloop in
future.

A related version of this change for b 1.8 is:
http://review.whamcloud.com/954

Change-Id: I63ebc7a8e0a8de6d34ffce5fd625df6db7216d32
Signed-off-by: Richard Henwood <rhenwood@whamcloud.com>