LU-409: only load llite_lloop module if kernel < 2.6.32 or LOAD_LLOOP
The change of block device interface in kernel 2.6.32-131 caused
the llite_lloop module to fail. llite_lloop module is currently
unsupported. This change introduces a test for kernel version
to prohibit loading a module with a kernel >= 2.6.32. A LOAD_LLOOP
flag (default false) is provided to ignore the kernel test and force
loading of llite_lloop if desired.

A related version of this change for b1_8 is:
http://review.whamcloud.com/954

Change-Id: I63ebc7a8e0a8de6d34ffce5fd625df6db7216d32
Signed-off-by: Richard Henwood <rhenwood@whamcloud.com>