LU-445 lnet: Send timestamps with LNet counters
We use a millisecond timer on a per-session basis. Each node then
sends the milliseconds since session start over with the LNet
counters. With 32bits and millisecond resolution, this should
allow ~49 days of session age before the timers would wrap.

This will change the wire attribute and not backward compatible.

Change-Id: I0dc6498d187ccc900b86d21b0460e1d419770033
Signed-off-by: Wally Wang <wang@cray.com>
Signed-off-by: Liang Zhen <liang@whamcloud.com>