LU-80 Large stripe support
Currently a file can be stripped across OSTs up to a limit of
160 stripes. Lustre 2.2 will expand this limit to 1350 stripes
or larger. Due to changes in the MDS LNET packet sizes interop
issues emerge with clients that are not aware of this feature.
This patch will allow Lustre 2.1 clients to function with
Lustre 2.1 servers without large stripe support and future
Lustre 2.2 servers that will have this feature.

Change-Id: Idbaeb98919aea6b4cd375b881ed87661034d9394
Signed-off-by: James Simmons <uja.ornl@gmail.com>