ORNL-1 skip unnecessary client obj lookup before insert
The client-side object will be initialized when inode is created in
memory. Such object will be inserted into global object hash table
for further using. The caller follows "lookup-alloc-lookup-insert".
It is standard, but maybe unnecessary for some cases. If it is sure
that there is only one caller will insert the new created object,
then we can skip the two lookup, and "alloc-insert" directly.

Change-Id: Ie173c57627b2e5b4ed9b8a93f368d88ba8e54c31
Signed-off-by: nasf <yong.fan@whamcloud.com>