LU-611 utils: add restripe option to lfs_migrate
Add the "-R" option to have lfs_migrate restripe a migrated file
instead of keeping the original striping.  This is useful of some
directory got the wrong striping and a bunch of files were created
with the wrong striping.

Avoid possible confusion between the lfs_migrate and lfs setstripe
command-line options.  For now, deprecate the old "-c" option, since
it is redundant in any case.

Change-Id: I3a39bad93ef5c079678c65960e53d22e51431df3
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
#
# total: 0 errors, 0 warnings, 96 lines checked
#
# Your patch has no obvious style problems and is ready for submission.
#
# total: 0 errors, 0 warnings, 96 lines checked
#
# Your patch has no obvious style problems and is ready for submission.
#
# total: 0 errors, 0 warnings, 96 lines checked
#
# Your patch has no obvious style problems and is ready for submission.
#
# total: 0 errors, 0 warnings, 96 lines checked
#
# Your patch has no obvious style problems and is ready for submission.
#
# total: 0 errors, 0 warnings, 96 lines checked
#
# Your patch has no obvious style problems and is ready for submission.