LU-670 Remove ldiskfs_dirhash from Lustre
The ldiskfs_dirhash() function is only used by the ldiskfs osd.
Since this functionality was originally copied from ldiskfs
it makes sense to remove it entirely and rely on the ldiskfs
implementation.

Normally this symbol isn't exported but since we are already
patching ldiskfs exporting it for use is trivial.  This is
currently done with the following ldiskfs patch for rhel6.

  ldiskfs/kernel_patches/patches/ext4-osd-iam-exports-rhel6.patch

Change-Id: I9a0ed25e4a2548328d91958608b6870fff7a8f88
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>