LU-553 build: Lustre style commit hooks for Git
Adds recommended commit hooks for Lustre commit messages, that
can/should be installed into lustre/.git/hooks for all Lustre trees.

prepare-commit-msg: adds the output of checkpatch.pl as a
comment to the commit message, to "soft" validate the new
patch is matching the Lustre coding style. It doesn't check
100% of the Lustre coding style, but at least finds most of
the obvious errors so that they can be fixed locally before
any manual inspection time is spent on the patch.

commit-msg: verifies the format of the commit message (somewhat).
Checks that there is a valid-looking Signed-off-by: line.
Adds the Gerrit ChangeId: line for change tracking.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I35fdfb8206d1f0efb3bb401522e8412c62b7c02a