TT-148 test: workaround for negative time shifts
Try to workaround negative test timestamps being seen in Maloo.
It appears that the negative time shifts are always even multiples
of 7h (-25200s) so try to compensate for this when printing out
test durations and the test completion summary line.

It doesn't just hard-code 25200s offset, since I suspect this is
somehow related to confusion by the client/host between UTC and
PDT (where the test clusters are hosted) and this offset will change
when daylight savings time arrives.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I95e6cc35986167dec86190d4f50d8b43dc9fabd3