ORNL-26 prevent call md_set_lock_data() repeatly
md_set_lock_data() is called from different functions in llite, and
may be called more than once for the same <inode lock> pair. It is
harmless for the correctness, but will affect the performance a bit,
should be avoided.

Drop dentry flags of "DCACHE_LUSTRE_INVALID" only when we hold the
"MDS_INODELOCK_LOOKUP" lock.

Change-Id: I4dbe206af77ba6a619f3268c238ce98ac7aef4c0
Signed-off-by: nasf <yong.fan@whamcloud.com>