LU-227 Call LB_PATH_LDISKFS then LB_CONFIG_MODULES
The LB_PATH_LDISKFS macro needs to be called prior to the
LB_CONFIG_MODULES. This is because the LB_CONFIG_MODULES calls
LC_PROG_LINUX which depends on the LDISKFS_DIR variable being set; and
this variable is set in the LB_PATH_LDISKFS macro.

As a result of the above, the location of the AC_DEFINE which provides
the HAVE_EXT4_LDISKFS macro had to be moved to a more appropriate
location (from LB_PATH_LDISKFS to LB_HAVE_EXT4_ENABLED). The move was
necessary because of the tightly coupled relationship between the
LDISKFS_DIR, LB_HAVE_EXT4_ENABLED, LB_PATH_LDISKFS macros.

To summarize the order of execution:
 1. LB_PATH_LDISKFS
 2. LDISKFS_DIR          (set in LB_PATH_LDISKFS)
 3. LB_CONFIG_MODULES
 4. LC_PROG_LINUX        (needs LDISKFS_DIR)
 5. LB_HAVE_EXT4_ENABLED (needs LDISKFS_DIR)
 6. HAVE_EXT4_LDISKFS    (set in LB_HAVE_EXT4_LDISKFS)

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Id22d38ac40b0f88c37379c35556c43536b4fe2f1