LU-227 Fix LC_EXT macros to be LDISKFS_DEVEL aware
Added a new AC_DEFINE call to define the LDISKFS_DEVEL macro. A few
LC_EXT* autoconf macros were also updated to use LDISKFS_DEVEL and 'do
the right thing' when building against the external ldiskfs devel
package. The specific autoconf macros that were updated are:

 * LC_EXT_INSERT_EXTENT_WITH_5ARGS
 * LC_EXT4_SINGLEDATA_TRANS_BLOCKS_SB
 * LC_EXT4_DISCARD_PREALLOCATIONS

These autoconf checks needed to be updated to use the ldiskfs headers
and symbols when the LDISKFS_DEVEL directive is set. Otherwise the
macros would always check against the ext headers and symbols, which
causes problems if full kernel sources are not available.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Icebbc794232d919ea0ef2e36bae40376ea5c14e9